refactor: macro-dashboard 컴포넌트 분해 (1,599줄 → 952줄)
- types.ts, utils.tsx, config.ts 분리
- 8개 컴포넌트를 components/ 폴더로 분리
- report-store addReport 타입 수정 (timestamp optional)
- 40.5% 코드 축소로 가독성/유지보수성 향상

Closes #7
