+++ components/features/macro/macro-dashboard.tsx
@@ -5,674 +5,27 @@ import {
     LineChart, Line, XAxis, YAxis, CartesianGrid, Tooltip,
     ResponsiveContainer, BarChart, Bar, ReferenceLine, Cell
 } from 'recharts';
-import { RefreshCw, TrendingUp, TrendingDown, Minus, CheckCircle, AlertTriangle, Sparkles, ChevronDown, ChevronUp, FileText } from 'lucide-react';
+import { RefreshCw, CheckCircle, AlertTriangle, Sparkles, ChevronDown, ChevronUp, FileText } from 'lucide-react';
+import type {
+    MacroScore,
+    GlobalMarketItem,
+    GlobalMarkets,
+    KrwUsdItem,
+    RiskIndicators,
+    MacroData,
+    EconomicEvent,
+    DailyReview
+} from './types';
+import { safetyBadge } from './utils';
+import { ZoneGauge } from './components/zone-gauge';
+import { LlmBriefingContent } from './components/llm-briefing-content';
+import { MarketCard } from './components/market-card';
+import { RiskCard } from './components/risk-card';
+import { ScoreCard } from './components/score-card';
+import { DailyReviewCard } from './components/daily-review-card';
+import { EconomicCalendarBanner } from './components/economic-calendar-banner';
+import { DateStrip } from './components/date-strip';
 
-// ─── Types ───────────────────────────────────────────────
-interface MacroScore {
-    value?: number | null;
-    change_pct?: number;
-    change_bps?: number;
-    score: number;
-    net_amount?: number;
-    net?: number;
-    volume?: number;
-}
-
-interface GlobalMarketItem {
-    value: number | null;
-    change_pct: number;
-}
-
-interface GlobalMarkets {
-    nasdaq: GlobalMarketItem;
-    sp500: GlobalMarketItem;
-    oil: GlobalMarketItem;
-    gold: GlobalMarketItem;
-    bitcoin: GlobalMarketItem;
-    krw_usd: KrwUsdItem;
-}
-
-interface KrwUsdItem {
-    value: number | null;
-    change: number;
-    change_pct: number;
-}
-
-interface RiskIndicators {
-    us2y?: { value?: number | null; change_bps?: number };
-    spread_2s10s?: number | null;
-    hy_spread?: { value?: number | null; change_bps?: number };
-    move?: { value?: number | null; change_pct?: number };
-    us30y?: { value?: number | null; change_bps?: number };
-}
-
-interface MacroData {
-    date: string;
-    kr_market_day?: string;
-    us_market_date?: string;
-    us_market_day?: string;
-    global_markets?: GlobalMarkets & { ewy?: GlobalMarketItem };
-    tier_a: {
-        dxy: MacroScore;
-        us10y: MacroScore;
-        vix: MacroScore;
-    };
-    tier_b: {
-        foreign_cash: MacroScore;
-        futures: MacroScore;
-        short_selling: MacroScore;
-    };
-    risk_indicators?: RiskIndicators;
-    overall_score: number;
-    safety_level: 1 | 2 | 3;
-    prediction_direction: string;
-    interpretation: string;
-    chain_analysis?: string;
-    llm_analysis?: string;
-    llm_analysis_pm?: string;
-    collected_at_am?: string;
-    collected_at_pm?: string;
-    news_context?: string;
-    economic_calendar?: {
-        economic_events: EconomicEvent[];
-        collected_at?: string;
-    };
-}
-
-interface EconomicEvent {
-    title: string;
-    date: string;
-    impact: 'High' | 'Medium';
-    forecast: string;
-    previous: string;
-}
-
-interface DailyReview {
-    date: string;
-    prediction: {
-        direction: string;
-        overall_score: number;
-        safety_level: number;
-        level_label: string;
-        summary: string;
-    };
-    actual: {
-        kospi_change: number;
-        foreign_net: number;
-        direction: string;
-    };
-    evaluation: {
-        hit: boolean;
-        label: string;
-        comment: string;
-    };
-}
-
-// ─── Score UI Helpers ────────────────────────────────────
-function scoreColor(score: number): string {
-    if (score <= -2) return 'text-red-400';
-    if (score === -1) return 'text-red-300';
-    if (score === 0) return 'text-slate-400';
-    if (score === 1) return 'text-emerald-300';
-    return 'text-emerald-400';
-}
-
-function scoreBg(score: number): string {
-    if (score <= -2) return 'bg-red-500/10 border-red-500/30';
-    if (score === -1) return 'bg-red-500/5 border-red-500/20';
-    if (score === 0) return 'bg-slate-800 border-slate-700';
-    if (score === 1) return 'bg-emerald-500/5 border-emerald-500/20';
-    return 'bg-emerald-500/10 border-emerald-500/30';
-}
-
-function scoreIcon(score: number) {
-    if (score > 0) return <TrendingUp className="w-3.5 h-3.5 text-emerald-400" />;
-    if (score < 0) return <TrendingDown className="w-3.5 h-3.5 text-red-400" />;
-    return <Minus className="w-3.5 h-3.5 text-slate-500" />;
-}
-
-function safetyBadge(level: number) {
-    const styles = {
-        1: 'bg-emerald-500/20 text-emerald-300 border-emerald-500/30',
-        2: 'bg-amber-500/20 text-amber-300 border-amber-500/30',
-        3: 'bg-red-500/20 text-red-300 border-red-500/30',
-    };
-    const labels = {
-        1: '🟢 Level 1: 외국인 우호적',
-        2: '🟡 Level 2: 중립/관망',
-        3: '🔴 Level 3: 외국인 이탈 경고',
-    };
-    return (
-        <span className={`inline-flex items-center px-3 py-1.5 rounded-lg text-sm font-bold border ${styles[level as keyof typeof styles] || styles[2]}`}>
-            {labels[level as keyof typeof labels] || labels[2]}
-        </span>
-    );
-}
-
-// ─── Zone Gauge (안전구간 시각화) ────────────────────────
-// 각 지표별 구간 정의: [매우위험, 위험, 중립, 안전, 매우안전] 경계값
-// position: 0(최악)~100(최선) 로 정규화
-type ZoneConfig = {
-    zones: { min: number; max: number; color: string; label: string }[];
-    normalize: (val: number) => number; // 값 → 0~100 위치
-    unit: string;
-    ticks: { pos: number; label: string }[];
-};
-
-const ZONE_CONFIGS: Record<string, ZoneConfig> = {
-    dxy_change: {
-        zones: [
-            { min: 0, max: 15, color: 'bg-red-500', label: '급등' },
-            { min: 15, max: 35, color: 'bg-red-400/60', label: '상승' },
-            { min: 35, max: 65, color: 'bg-slate-500', label: '보합' },
-            { min: 65, max: 85, color: 'bg-emerald-400/60', label: '하락' },
-            { min: 85, max: 100, color: 'bg-emerald-500', label: '급락' },
-        ],
-        normalize: (v) => Math.max(0, Math.min(100, 50 - v * 25)),  // +2%→0, 0→50, -2%→100
-        unit: '%',
-        ticks: [{ pos: 15, label: '+1.5%' }, { pos: 35, label: '+0.5%' }, { pos: 65, label: '-0.5%' }, { pos: 85, label: '-1.5%' }],
-    },
-    us10y_change: {
-        zones: [
-            { min: 0, max: 15, color: 'bg-red-500', label: '급등' },
-            { min: 15, max: 35, color: 'bg-red-400/60', label: '상승' },
-            { min: 35, max: 65, color: 'bg-slate-500', label: '보합' },
-            { min: 65, max: 85, color: 'bg-emerald-400/60', label: '하락' },
-            { min: 85, max: 100, color: 'bg-emerald-500', label: '급락' },
-        ],
-        normalize: (v) => Math.max(0, Math.min(100, 50 - v * 3.33)),  // +15bp→0, 0→50, -15bp→100
-        unit: 'bp',
-        ticks: [{ pos: 15, label: '+10bp' }, { pos: 35, label: '+3bp' }, { pos: 65, label: '-3bp' }, { pos: 85, label: '-10bp' }],
-    },
-    vix_level: {
-        zones: [
-            { min: 0, max: 15, color: 'bg-red-500', label: '패닉' },
-            { min: 15, max: 30, color: 'bg-red-400/60', label: '불안' },
-            { min: 30, max: 60, color: 'bg-slate-500', label: '정상' },
-            { min: 60, max: 80, color: 'bg-emerald-400/60', label: '안정' },
-            { min: 80, max: 100, color: 'bg-emerald-500', label: '낙관' },
-        ],
-        normalize: (v) => Math.max(0, Math.min(100, (35 - v) * 2.86)),  // VIX 35→0, 20→43, 10→71, 0→100
-        unit: '',
-        ticks: [{ pos: 15, label: '30' }, { pos: 30, label: '25' }, { pos: 60, label: '15' }, { pos: 80, label: '12' }],
-    },
-    foreign: {
-        zones: [
-            { min: 0, max: 10, color: 'bg-red-500', label: '대량매도' },
-            { min: 10, max: 35, color: 'bg-red-400/60', label: '매도' },
-            { min: 35, max: 65, color: 'bg-slate-500', label: '중립' },
-            { min: 65, max: 90, color: 'bg-emerald-400/60', label: '매수' },
-            { min: 90, max: 100, color: 'bg-emerald-500', label: '대량매수' },
-        ],
-        normalize: (v) => Math.max(0, Math.min(100, 50 + v / 100)),  // -5000→0, 0→50, +5000→100
-        unit: '억',
-        ticks: [{ pos: 10, label: '-5천' }, { pos: 35, label: '-500' }, { pos: 65, label: '+500' }, { pos: 90, label: '+5천' }],
-    },
-    futures: {
-        zones: [
-            { min: 0, max: 10, color: 'bg-red-500', label: '대량매도' },
-            { min: 10, max: 35, color: 'bg-red-400/60', label: '매도' },
-            { min: 35, max: 65, color: 'bg-slate-500', label: '중립' },
-            { min: 65, max: 90, color: 'bg-emerald-400/60', label: '매수' },
-            { min: 90, max: 100, color: 'bg-emerald-500', label: '대량매수' },
-        ],
-        normalize: (v) => Math.max(0, Math.min(100, 50 + v / 250)),  // -12500→0, 0→50, +12500→100
-        unit: '계약',
-        ticks: [{ pos: 10, label: '-1만' }, { pos: 35, label: '-2천' }, { pos: 65, label: '+2천' }, { pos: 90, label: '+1만' }],
-    },
-    short_change: {
-        zones: [
-            { min: 0, max: 10, color: 'bg-red-500', label: '급증' },
-            { min: 10, max: 35, color: 'bg-red-400/60', label: '증가' },
-            { min: 35, max: 65, color: 'bg-slate-500', label: '보합' },
-            { min: 65, max: 90, color: 'bg-emerald-400/60', label: '감소' },
-            { min: 90, max: 100, color: 'bg-emerald-500', label: '급감' },
-        ],
-        normalize: (v) => Math.max(0, Math.min(100, 50 - v * 1.25)),  // +40→0, 0→50, -40→100
-        unit: '%',
-        ticks: [{ pos: 10, label: '+30%' }, { pos: 35, label: '+10%' }, { pos: 65, label: '-10%' }, { pos: 90, label: '-30%' }],
-    },
-};
-
-function ZoneGauge({ zoneKey, rawValue }: { zoneKey: string; rawValue: number }) {
-    const config = ZONE_CONFIGS[zoneKey];
-    if (!config) return null;
-
-    const pos = config.normalize(rawValue);
-    // 현재 위치가 어느 zone에 있는지
-    const currentZone = config.zones.find(z => pos >= z.min && pos < z.max) || config.zones[2];
-
-    return (
-        <div className="mt-2.5 pt-2 border-t border-slate-700/50">
-            {/* Zone label */}
-            <div className="flex items-center justify-between mb-1.5">
-                <span className="text-[10px] text-slate-500">안전 구간</span>
-                <span className={`text-[10px] font-bold ${
-                    currentZone.color.includes('red') ? 'text-red-400' :
-                    currentZone.color.includes('emerald') ? 'text-emerald-400' : 'text-slate-400'
-                }`}>
-                    {currentZone.label}
-                </span>
-            </div>
-            {/* 5-zone bar */}
-            <div className="relative w-full h-2 rounded-full flex overflow-hidden">
-                {config.zones.map((zone, i) => (
-                    <div
-                        key={i}
-                        className={`h-full ${zone.color} ${i === 0 ? 'rounded-l-full' : ''} ${i === config.zones.length - 1 ? 'rounded-r-full' : ''}`}
-                        style={{ width: `${zone.max - zone.min}%` }}
-                    />
-                ))}
-                {/* Needle / indicator */}
-                <div
-                    className="absolute top-[-2px] w-[3px] h-[12px] bg-white rounded-full shadow-[0_0_4px_rgba(255,255,255,0.8)]"
-                    style={{ left: `calc(${pos}% - 1.5px)` }}
-                />
-            </div>
-            {/* Tick labels */}
-            <div className="relative w-full h-3 mt-0.5">
-                {config.ticks.map((tick, i) => (
-                    <span
-                        key={i}
-                        className="absolute text-[8px] text-slate-600 -translate-x-1/2"
-                        style={{ left: `${tick.pos}%` }}
-                    >
-                        {tick.label}
-                    </span>
-                ))}
-            </div>
-        </div>
-    );
-}
-
-// ─── LLM Briefing Renderer ───────────────────────────────
-function formatTextParts(text: string) {
-    const parts = text.split(/(\*\*.*?\*\*)/g);
-    return parts.map((part, i) => {
-        if (part.startsWith('**') && part.endsWith('**')) {
-            return <strong key={i} className="text-white font-semibold">{part.slice(2, -2)}</strong>;
-        }
-        return <span key={i}>{part}</span>;
-    });
-}
-
-function LlmBriefingContent({ content }: { content: string }) {
-    return (
-        <div className="space-y-1">
-            {content.split('\n').map((line, i) => {
-                const trimmed = line.trim();
-                if (!trimmed) return <div key={i} className="h-2" />;
-
-                if (trimmed.startsWith('### ')) {
-                    return (
-                        <h4 key={i} className="text-purple-300 font-bold mt-4 mb-1 text-sm flex items-center gap-2">
-                            {formatTextParts(trimmed.slice(4))}
-                        </h4>
-                    );
-                }
-                if (trimmed.startsWith('## ')) {
-                    return (
-                        <h3 key={i} className="text-purple-200 font-bold mt-5 mb-2 text-base">
-                            {formatTextParts(trimmed.slice(3))}
-                        </h3>
-                    );
-                }
-                if (trimmed.startsWith('# ')) {
-                    return (
-                        <h2 key={i} className="text-purple-100 font-bold mt-5 mb-2 text-lg">
-                            {formatTextParts(trimmed.slice(2))}
-                        </h2>
-                    );
-                }
-                if (trimmed.startsWith('- ') || trimmed.startsWith('* ')) {
-                    return (
-                        <div key={i} className="flex gap-2 ml-2">
-                            <span className="text-purple-400 shrink-0">·</span>
-                            <span className="text-slate-300 text-sm">{formatTextParts(trimmed.slice(2))}</span>
-                        </div>
-                    );
-                }
-                if (/^\d+\.\s/.test(trimmed)) {
-                    const match = trimmed.match(/^(\d+\.)\s(.*)$/);
-                    return (
-                        <div key={i} className="flex gap-2 ml-2">
-                            <span className="text-purple-400 shrink-0 text-xs font-mono">{match?.[1]}</span>
-                            <span className="text-slate-300 text-sm">{formatTextParts(match?.[2] || trimmed)}</span>
-                        </div>
-                    );
-                }
-                if (trimmed.startsWith('> ')) {
-                    return (
-                        <div key={i} className="border-l-2 border-purple-500/50 pl-3 ml-2 text-slate-400 text-sm italic">
-                            {formatTextParts(trimmed.slice(2))}
-                        </div>
-                    );
-                }
-                if (trimmed.startsWith('---') || trimmed.startsWith('***')) {
-                    return <hr key={i} className="border-slate-700/50 my-3" />;
-                }
-                return (
-                    <p key={i} className="text-slate-300 text-sm leading-relaxed">
-                        {formatTextParts(trimmed)}
-                    </p>
-                );
-            })}
-        </div>
-    );
-}
-
-// ─── Market Card (Global Markets) ────────────────────────
-function MarketCard({ label, icon, value, changePct, unit }: {
-    label: string;
-    icon: string;
-    value: string;
-    changePct: number;
-    unit?: string;
-}) {
-    const isPositive = changePct > 0;
-    const isNegative = changePct < 0;
-    const colorClass = isPositive ? 'text-emerald-400' : isNegative ? 'text-red-400' : 'text-slate-400';
-    const bgClass = isPositive ? 'bg-emerald-500/5 border-emerald-500/20' : isNegative ? 'bg-red-500/5 border-red-500/20' : 'bg-slate-800 border-slate-700';
-
-    return (
-        <div className={`rounded-lg border p-3 ${bgClass}`}>
-            <div className="text-lg mb-1">{icon}</div>
-            <div className="text-[10px] text-slate-500 mb-1">{label}</div>
-            <div className="text-sm font-bold text-white">{value}{unit || ''}</div>
-            <div className={`text-xs font-mono mt-0.5 ${colorClass}`}>
-                {changePct >= 0 ? '+' : ''}{changePct.toFixed(2)}%
-            </div>
-        </div>
-    );
-}
-
-// ─── Risk Indicator Card (no scoring) ────────────────────
-function RiskCard({ label, value, subValue, subColor, desc }: {
-    label: string;
-    value: string;
-    subValue: string;
-    subColor?: 'red' | 'green' | 'neutral';
-    desc?: string;
-}) {
-    const color = subColor === 'red' ? 'text-red-400' : subColor === 'green' ? 'text-emerald-400' : 'text-slate-400';
-    return (
-        <div className="rounded-lg border border-slate-700/50 bg-slate-900/30 p-3">
-            <div className="text-[10px] text-slate-500 mb-0.5">{label}</div>
-            {desc && <div className="text-[9px] text-slate-600 mb-1">{desc}</div>}
-            <div className="text-sm font-bold text-white">{value}</div>
-            <div className={`text-xs mt-0.5 ${color}`}>{subValue}</div>
-        </div>
-    );
-}
-
-// ─── Score Card ──────────────────────────────────────────
-function ScoreCard({ label, icon, value, subValue, unit, score, accuracy, zoneKey, rawValue }: {
-    label: string;
-    icon: string;
-    value: string;
-    subValue: string;
-    unit?: string;
-    score: number;
-    accuracy?: number;
-    zoneKey?: string;
-    rawValue?: number;
-}) {
-    return (
-        <div className={`rounded-lg border p-4 ${scoreBg(score)}`}>
-            <div className="flex items-center justify-between mb-2">
-                <span className="text-lg">{icon}</span>
-                <div className="flex items-center gap-1">
-                    {scoreIcon(score)}
-                    <span className={`text-sm font-bold ${scoreColor(score)}`}>{score > 0 ? '+' : ''}{score}</span>
-                </div>
-            </div>
-            <div className="text-xs text-slate-500 mb-1">{label}</div>
-            <div className="text-lg font-bold text-white">{value}</div>
-            <div className="text-xs text-slate-400 mt-0.5">{subValue}{unit}</div>
-
-            {/* Zone Gauge */}
-            {zoneKey && rawValue !== undefined && <ZoneGauge zoneKey={zoneKey} rawValue={rawValue} />}
-
-            {accuracy !== undefined && (
-                <div className="mt-2 pt-2 border-t border-slate-700/50">
-                    <div className="flex items-center justify-between">
-                        <span className="text-[10px] text-slate-500">적중률</span>
-                        <span className="text-xs font-medium text-slate-300">{(accuracy * 100).toFixed(0)}%</span>
-                    </div>
-                    <div className="w-full bg-slate-700 rounded-full h-1 mt-1">
-                        <div
-                            className={`h-1 rounded-full ${accuracy >= 0.7 ? 'bg-emerald-400' : accuracy >= 0.5 ? 'bg-amber-400' : 'bg-red-400'}`}
-                            style={{ width: `${accuracy * 100}%` }}
-                        />
-                    </div>
-                </div>
-            )}
-        </div>
-    );
-}
-
-// ─── Daily Review Card ───────────────────────────────────
-function DailyReviewCard({ review }: { review: DailyReview }) {
-    return (
-        <div className="rounded-lg border border-slate-700 bg-slate-900/50 p-5 mb-6">
-            <div className="flex items-center gap-2 mb-3">
-                <span className="text-base font-bold text-white">📋 전일 리뷰</span>
-                <span className="text-xs text-slate-500">({review.date})</span>
-                <span className={`ml-auto px-2 py-0.5 rounded text-xs font-bold ${review.evaluation.hit ? 'bg-emerald-500/20 text-emerald-300' : 'bg-amber-500/20 text-amber-300'}`}>
-                    {review.evaluation.label}
-                </span>
-            </div>
-
-            <div className="grid grid-cols-2 gap-4">
-                {/* Prediction */}
-                <div>
-                    <div className="text-xs text-slate-500 mb-1">예측</div>
-                    <div className="flex items-center gap-2 mb-1">
-                        <span className="text-sm font-medium text-slate-200">
-                            {review.prediction.level_label}
-                        </span>
-                        <span className="text-xs text-slate-400">
-                            ({review.prediction.overall_score >= 0 ? '+' : ''}{review.prediction.overall_score.toFixed(2)})
-                        </span>
-                    </div>
-                    <p className="text-xs text-slate-400 leading-relaxed">
-                        {review.prediction.summary}
-                    </p>
-                </div>
-
-                {/* Actual */}
-                <div>
-                    <div className="text-xs text-slate-500 mb-1">실제</div>
-                    <div className="flex items-center gap-3 mb-1">
-                        <span className={`text-sm font-bold ${review.actual.kospi_change >= 0 ? 'text-emerald-400' : 'text-red-400'}`}>
-                            KOSPI {review.actual.kospi_change >= 0 ? '+' : ''}{review.actual.kospi_change.toFixed(1)}%
-                        </span>
-                        <span className="text-xs text-slate-400">
-                            외국인 {review.actual.foreign_net >= 0 ? '+' : ''}{review.actual.foreign_net.toLocaleString()}억
-                        </span>
-                    </div>
-                </div>
-            </div>
-
-            {/* Evaluation */}
-            <div className="mt-3 pt-3 border-t border-slate-800">
-                <div className="flex items-start gap-2">
-                    {review.evaluation.hit
-                        ? <CheckCircle className="w-4 h-4 text-emerald-400 mt-0.5 shrink-0" />
-                        : <AlertTriangle className="w-4 h-4 text-amber-400 mt-0.5 shrink-0" />
-                    }
-                    <p className="text-xs text-slate-300 leading-relaxed">{review.evaluation.comment}</p>
-                </div>
-            </div>
-        </div>
-    );
-}
-
-// ─── Economic Calendar Banner (공지사항 스타일) ─────────────
-function EconomicCalendarBanner({ events }: { events: EconomicEvent[] }) {
-    const [expanded, setExpanded] = useState(false);
-    const today = new Date().toISOString().slice(0, 10);
-
-    // 오늘 이벤트
-    const todayEvents = events.filter(ev => ev.date?.slice(0, 10) === today);
-    // 이번 주 High 이벤트 (오늘 제외)
-    const upcomingHigh = events.filter(ev => ev.date?.slice(0, 10) !== today && ev.impact === 'High');
-
-    // 헤더 요약 텍스트
-    const highCount = events.filter(e => e.impact === 'High').length;
-    const summary = todayEvents.length > 0
-        ? `오늘 ${todayEvents.length}건 — ${todayEvents.map(e => e.title).join(', ')}`
-        : `이번 주 주요 지표 ${highCount}건`;
-
-    return (
-        <div className="rounded-lg border border-indigo-500/30 bg-indigo-950/20">
-            {/* 배너 헤더 — 항상 보임 */}
-            <button
-                onClick={() => setExpanded(!expanded)}
-                className="w-full flex items-center gap-2 px-4 py-2.5 text-left hover:bg-indigo-950/30 transition-colors rounded-lg"
-            >
-                <span className="text-sm">📅</span>
-                <span className="text-xs font-medium text-indigo-300 flex-1 truncate">{summary}</span>
-                {highCount > 0 && (
-                    <span className="shrink-0 px-1.5 py-0.5 rounded text-[10px] font-bold bg-red-500/20 text-red-300">
-                        HIGH {highCount}
-                    </span>
-                )}
-                {expanded
-                    ? <ChevronUp className="w-3.5 h-3.5 text-slate-500 shrink-0" />
-                    : <ChevronDown className="w-3.5 h-3.5 text-slate-500 shrink-0" />
-                }
-            </button>
-
-            {/* 펼친 상세 */}
-            {expanded && (
-                <div className="px-4 pb-3 space-y-2">
-                    {(() => {
-                        const grouped: Record<string, EconomicEvent[]> = {};
-                        for (const ev of events) {
-                            const dk = ev.date?.slice(0, 10) || '';
-                            if (!grouped[dk]) grouped[dk] = [];
-                            grouped[dk].push(ev);
-                        }
-                        return Object.keys(grouped).sort().map(dateKey => {
-                            const d = new Date(dateKey + 'T00:00:00');
-                            const dayNames = ['일', '월', '화', '수', '목', '금', '토'];
-                            const label = `${d.getMonth() + 1}/${d.getDate()} (${dayNames[d.getDay()]})`;
-                            const isToday = dateKey === today;
-
-                            return (
-                                <div key={dateKey}>
-                                    <div className={`text-[10px] font-semibold mb-1 ${isToday ? 'text-blue-400' : 'text-slate-500'}`}>
-                                        {label} {isToday && '  TODAY'}
-                                    </div>
-                                    {grouped[dateKey].map((ev, i) => {
-                                        const time = ev.date?.slice(11, 16) || '';
-                                        return (
-                                            <div key={i} className="flex items-center gap-2 py-0.5 text-[11px]">
-                                                <span className={`w-1.5 h-1.5 rounded-full shrink-0 ${ev.impact === 'High' ? 'bg-red-400' : 'bg-amber-400'}`} />
-                                                <span className="text-slate-500 font-mono w-10 shrink-0">{time}</span>
-                                                <span className="text-slate-300 flex-1">{ev.title}</span>
-                                                {ev.forecast && (
-                                                    <span className="text-slate-500 shrink-0 text-[10px]">
-                                                        {ev.forecast} / {ev.previous}
-                                                    </span>
-                                                )}
-                                            </div>
-                                        );
-                                    })}
-                                </div>
-                            );
-                        });
-                    })()}
-                </div>
-            )}
-        </div>
-    );
-}
-
-// ─── Date Strip (가로 날짜 네비게이션) ─────────────────────
-const DAY_NAMES = ['일', '월', '화', '수', '목', '금', '토'];
-
-function DateStrip({
-    availableDates, selectedDate, onSelectDate
-}: {
-    availableDates: {date: string; score: number; safety_level: number; has_llm: boolean}[];
-    selectedDate: string | null;
-    onSelectDate: (date: string | null) => void;
-}) {
-    const toLocal = (d: Date) => `${d.getFullYear()}-${String(d.getMonth()+1).padStart(2,'0')}-${String(d.getDate()).padStart(2,'0')}`;
-    const today = toLocal(new Date());
-    const dateMap = new Map(availableDates.map(d => [d.date, d]));
-
-    // 최근 14일 생성 (오늘 포함)
-    const days: string[] = [];
-    for (let i = 13; i >= 0; i--) {
-        const d = new Date();
-        d.setDate(d.getDate() - i);
-        days.push(toLocal(d));
-    }
-
-    return (
-        <div className="flex items-center gap-1.5 overflow-x-auto pb-1 scrollbar-hide">
-            {days.map((dateStr) => {
-                const info = dateMap.get(dateStr);
-                const isToday = dateStr === today;
-                const isSelected = dateStr === selectedDate || (selectedDate === null && isToday);
-                const hasData = !!info;
-                const hasLlm = info?.has_llm || false;
-
-                const d = new Date(dateStr + 'T00:00:00');
-                const dayNum = d.getDate();
-                const dayName = DAY_NAMES[d.getDay()];
-                const isWeekend = d.getDay() === 0 || d.getDay() === 6;
-
-                const safetyBorder = info
-                    ? info.safety_level === 1 ? 'border-emerald-500/60' : info.safety_level === 3 ? 'border-red-500/60' : 'border-amber-500/60'
-                    : 'border-slate-700/50';
-
-                return (
-                    <button
-                        key={dateStr}
-                        onClick={() => {
-                            if (isToday) onSelectDate(null);
-                            else onSelectDate(dateStr);
-                        }}
-                        className={`relative flex flex-col items-center shrink-0 w-12 py-1.5 rounded-lg border transition-all cursor-pointer
-                            ${isSelected
-                                ? 'bg-blue-600/20 border-blue-500 text-white'
-                                : hasData
-                                    ? `bg-slate-800/50 ${safetyBorder} text-slate-300 hover:bg-slate-700/50`
-                                    : 'bg-slate-900/20 border-slate-800/30 text-slate-600 hover:bg-slate-800/30'
-                            }
-                        `}
-                    >
-                        <span className={`text-[9px] leading-none ${isWeekend && !isSelected ? 'text-slate-600' : ''}`}>
-                            {dayName}
-                        </span>
-                        <span className={`text-sm font-bold leading-tight mt-0.5 ${isSelected ? 'text-white' : ''}`}>
-                            {dayNum}
-                        </span>
-                        {/* 하단 인디케이터 */}
-                        <div className="flex items-center gap-0.5 mt-0.5 h-1.5">
-                            {hasData && (
-                                <span className={`w-1 h-1 rounded-full ${
-                                    info!.safety_level === 1 ? 'bg-emerald-400' :
-                                    info!.safety_level === 3 ? 'bg-red-400' : 'bg-amber-400'
-                                }`} />
-                            )}
-                            {hasLlm && (
-                                <span className="w-1 h-1 rounded-full bg-purple-400" />
-                            )}
-                        </div>
-                        {isToday && !isSelected && (
-                            <span className="absolute -top-0.5 left-1/2 -translate-x-1/2 w-1 h-1 rounded-full bg-blue-400" />
-                        )}
-                    </button>
-                );
-            })}
-        </div>
-    );
-}
 
 // ─── Main Component ──────────────────────────────────────
 export function MacroDashboard() {
